refactor: Rename `…/event-detection/event-types/apiGatewayV2.js`
