+++ sdk-js/src/lib/event-detection/event-types/api-gateway-v2.js
