+++ sdk-js/src/lib/event-detection/index.js
@@ -2,7 +2,7 @@
 
 const alexaSkill = require('./event-types/alexaSkill');
 const apiGateway = require('./event-types/apiGateway');
-const apiGatewayV2 = require('./event-types/apiGatewayV2');
+const apiGatewayV2 = require('./event-types/api-gateway-v2');
 const customAuthorizer = require('./event-types/customAuthorizer');
 const cloudFront = require('./event-types/cloudFront');
 const cloudwatchEvent = require('./event-types/cloudwatchEvent');
